refactor: Atomic Design Pattern에 따라 Image/Input 컴포넌트 경로 조정
