+++ src/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Category from 'components/templates/Category';
 import Item from 'components/templates/Item';
-import ProductOptionImage from 'components/organisms/ProductOption/Image';
-import ProductOptionInput from 'components/organisms/ProductOption/Input';
+import ProductOptionImage from 'components/atoms/ImageUploader';
+import ProductOptionInput from 'components/organisms/ProductOptionInput';
 
 function App() {
   return (
@@ -24,6 +24,7 @@ function App() {
           <div>hello</div>
         </Item>
       </Category>
+
       <Category title="상품 옵션*">
         <ProductOptionImage />
         <ProductOptionInput />
