+++ src/components/atoms/ImageUploader/index.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { ImageWrapper } from './style';
+import { UploaderWrapper } from './style';
 
-export default function ProductOptionImage() {
+export default function ImageUploader() {
   return (
-    <ImageWrapper>
+    <UploaderWrapper>
       {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
       <label htmlFor="product-image">+ 이미지 첨부</label>
       <input id="product-image" type="file" accept="image/*" />
-    </ImageWrapper>
+    </UploaderWrapper>
   );
 }
